+++ ale/spiceql_access.py
@@ -90,6 +90,19 @@ def spiceql_call(function_name = "", function_args = {}, use_web=False):
     logger.debug(f"Calling {function_name} with args: {function_args}")
     if use_web == False:
         function_args["useWeb"] = False
+        ephem_funcs = ["getTargetStates", "getTargetOrientations"]
+        if function_name in ephem_funcs and "ets" not in function_args.keys():
+            if ("startEt" in function_args.keys() and 
+                "stopEt" in function_args.keys() and 
+                "numRecords" in function_args.keys()):
+                start = function_args.pop("startEt")
+                stop = function_args.pop("stopEt")
+                num_records = function_args.pop("numRecords")
+                ets = np.linspace(start, stop, num_records)
+                function_args['ets'] = ets
+            else:
+                raise RuntimeError(f"No ephemeris data provided, {function_name} needs 'ets' or " + 
+                                    "'startEt', 'stopet', 'numRecords' in `function_args`")
         func = getattr(pyspiceql, function_name)
         ret = func(**function_args)[0]
         return ret
@@ -109,13 +122,19 @@ def spiceql_call(function_name = "", function_args = {}, use_web=False):
     clean_function_args = stringify_web_args(function_args)
     logger.debug("Args: " + str(clean_function_args))
 
-    if function_name == "getTargetStates":
+    # if it's using ets lists, use post 
+    if function_name == "getTargetStates" and "ets" in function_args.keys():
         post_body = str(clean_function_args).replace("\'", "\"")
         logger.debug("getTargetStates POST Payload: " + post_body)
         response = requests.post(url, data=post_body, headers=headers, verify=False)
     else:
         response = requests.get(url, params=clean_function_args, headers=headers, verify=False)
-    check_response(response)
+
+    response.raise_for_status(); 
+    if response.status_code != 200:
+        raise requests.HTTPError(f"{response.url} Received code {response.status_code} from spice server, with error: {response.json()}")
+    if response.json()["statusCode"] != 200:
+        raise requests.HTTPError(f"{response.url} Received code {response.json()['statusCode']} from spice server, with error: {response.json()['body']}")
 
     logger.debug(f"Request URL={str(response.url)}")
     logger.debug(f"Kernels={str(response.json()['body']['kernels'])}")
@@ -160,6 +179,7 @@ def get_ephem_data(times, function_name, batch_size=300, web=False, function_arg
               requested in times.
     """
     valid_functions = ["getTargetOrientations", "getTargetStates"]
+    
     if function_name not in valid_functions:
         raise ValueError(f"The function name {function_name} is not supported "
                           "by this function please pass one of the following: " + str(valid_functions))